info box cleanup
