+++ src/blocks/info-box/components/IconImage.js
@@ -4,11 +4,9 @@ const {
 
 const { __ } = wp.i18n
 
-class InfoBoxIconImage extends React.Component {
+export default function InfoBoxIconImage(props){
 
-	render() {
-
-		const { attributes } = this.props 
+		const { attributes } = props 
 
 		let url_chk = ""
 		if( typeof attributes.iconImage !== "undefined" && attributes.iconImage !== null && attributes.iconImage !=="" ){
@@ -43,7 +41,5 @@ class InfoBoxIconImage extends React.Component {
 			return null
 		}       
 		
-	}
 }
 
-export default InfoBoxIconImage
