+++ src/blocks/info-box/components/Title.js
@@ -2,22 +2,15 @@ const {
 	RichText,
 } = wp.blockEditor
 
-const {
-	createBlock
-} = wp.blocks
-
 const { __ } = wp.i18n
 
-class Title extends React.Component {
-
-	render() {
-
+export default function InfoBoxRender(props){
+		
 		//const { attributes, setAttributes , props } = this.props;
 		const {
 			attributes,
 			setAttributes ,
-			props
-		} = this.props
+		} = props
 
 		if( setAttributes !== "not_set" ){
 			return (
@@ -28,20 +21,7 @@ class Title extends React.Component {
 	                className = 'uagb-ifb-title'
 	                onChange = { ( value ) => setAttributes( { infoBoxTitle: value } ) }
 	                multiline={ false }
-	                onMerge = { props.mergeBlocks }
-	                onSplit = {
-						props.insertBlocksAfter ?
-							( before, after, ...blocks ) => {
-								setAttributes( { content: before } )
-								props.insertBlocksAfter( [
-									...blocks,
-									createBlock( "core/paragraph", { content: after } ),
-								] )
-							} :
-							undefined
-					}
-					onRemove={ () => props.onReplace( [] ) }
-	            />
+	             />
 			)
 		}else{
 			return (
@@ -52,7 +32,6 @@ class Title extends React.Component {
 	            />
 			)
 		}
-	}
+
 }
 
-export default Title
